Refactor Login Page
